// models/letter-types.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lob
 * The Lob API is organized around REST. Our API is designed to have predictable, resource-oriented URLs and uses HTTP response codes to indicate any API errors. <p> Looking for our [previous documentation](https://lob.github.io/legacy-docs/)? 
 *
 * The version of the OpenAPI document: 1.3.0
 * Contact: lob-openapi@lob.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Unique identifier referring to status of letter
 * @export
 * @enum {string}
 */

export enum LetterTypes {
    Created = 'letter.created',
    RenderedPdf = 'letter.rendered_pdf',
    RenderedThumbnails = 'letter.rendered_thumbnails',
    Deleted = 'letter.deleted',
    Mailed = 'letter.mailed',
    InTransit = 'letter.in_transit',
    InLocalArea = 'letter.in_local_area',
    ProcessedForDelivery = 'letter.processed_for_delivery',
    ReRouted = 'letter.re-routed',
    ReturnedToSender = 'letter.returned_to_sender',
    CertifiedMailed = 'letter.certified.mailed',
    CertifiedInTransit = 'letter.certified.in_transit',
    CertifiedInLocalArea = 'letter.certified.in_local_area',
    CertifiedProcessedForDelivery = 'letter.certified.processed_for_delivery',
    CertifiedReRouted = 'letter.certified.re-routed',
    CertifiedReturnedToSender = 'letter.certified.returned_to_sender',
    CertifiedDelivered = 'letter.certified.delivered',
    CertifiedPickupAvailable = 'letter.certified.pickup_available',
    CertifiedIssue = 'letter.certified.issue',
    ReturnEnvelopeCreated = 'letter.return_envelope.created',
    ReturnEnvelopeInTransit = 'letter.return_envelope.in_transit',
    ReturnEnvelopeInLocalArea = 'letter.return_envelope.in_local_area',
    ReturnEnvelopeProcessedForDelivery = 'letter.return_envelope.processed_for_delivery',
    ReturnEnvelopeReRouted = 'letter.return_envelope.re_routed',
    ReturnEnvelopeReturnedToSender = 'letter.return_envelope.returned_to_sender'
}



/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */